[P4ADEV-2982] Refactor routing structure and setup logic
- Moved routing definitions from App.tsx to a new index.tsx file in the routes directory for better organization.
- Introduced a setup function to handle application initialization and interceptor configuration.
- Updated imports across various components to reference the new routing structure.
- Simplified the App component by removing direct route definitions and utilizing the new router.
- Enhanced error handling and loading states in the RouterProvider.
