+++ src/components/ReportingSearchResults/ReportingDataGrid.tsx
@@ -9,7 +9,7 @@ import ActionMenu from '../ActionMenu/ActionMenu';
 import CustomDataGrid from '../DataGrid/CustomDataGrid';
 import { FileDownload, Visibility } from '@mui/icons-material';
 import { generatePath, useNavigate } from 'react-router-dom';
-import { PageRoutes } from '../../App';
+import { PageRoutes } from '../../routes';
 import { PagedPaymentsReportingView } from '../../../generated/data-contracts';
 import { moneyFormat } from '../../utils/formatters';
