+++ src/components/Treasury/Treasury.tsx
@@ -4,7 +4,7 @@ import { FileUpload } from '@mui/icons-material';
 import { Grid } from '@mui/material';
 import { useTranslation } from 'react-i18next';
 import TitleComponent from '../TitleComponent/TitleComponent';
-import { PageRoutes } from '../../App';
+import { PageRoutes } from '../../routes';
 import { generatePath, useNavigate } from 'react-router';
 import { useMultiFilters } from '../../hooks/useMultiFilters';
