+++ src/components/TreasurySearchResults/TreasurySearchResults.tsx
@@ -6,7 +6,7 @@ import { ButtonNaked } from '@pagopa/mui-italia';
 import { FilterAlt } from '@mui/icons-material';
 import { useState } from 'react';
 import { generatePath, useNavigate } from 'react-router';
-import { PageRoutes } from '../../App';
+import { PageRoutes } from '../../routes';
 import { FilterMap, useMultiFilters } from '../../hooks/useMultiFilters';
 import { FilterDrawer } from '../Drawer/FilterDrawer';
 import { BaseFilterValues } from '../../models/Filters';
