+++ src/routes/DebtPositionDetail/components/PaymentOptionSection.test.tsx
@@ -4,7 +4,7 @@ import { describe, it, expect, beforeEach, vi, afterEach } from 'vitest';
 import { i18nTestSetup } from '../../../__tests__/i18nTestSetup';
 import { render } from '../../../__tests__/renderers';
 import { PaymentOptionDisplayData } from '../DebtPositionDetail';
-import { PageRoutes } from '../../../App';
+import { PageRoutes } from '../../../routes';
 
 const navigateMock = vi.fn();
