+++ src/routes/DebtPositions/components/DebtPositionsDataGrid.tsx
@@ -7,7 +7,7 @@ import {
 import { useTranslation } from 'react-i18next';
 import { ChevronRight } from '@mui/icons-material';
 import { generatePath, useNavigate } from 'react-router-dom';
-import { PageRoutes } from '../../../App';
+import { PageRoutes } from '../../../routes';
 import CustomDataGrid from '../../../components/DataGrid/CustomDataGrid';
 import Chip, { ChipProps } from '@mui/material/Chip';
 import {
