+++ src/routes/DebtTypeOrgCreate/index.tsx
@@ -3,7 +3,7 @@ import { useTranslation } from 'react-i18next';
 import { Stepper } from '../../components/Stepper/types';
 import { StepperContainer } from '../../components/Stepper';
 import { useNavigate } from 'react-router';
-import { PageRoutes } from '../../App';
+import { PageRoutes } from '../../routes';
 import { Step1Configuration } from './steps/Step1Configuration';
 import { Step2Behaviour } from './steps/Step2Behaviour';
 import { Step3Accounting } from './steps/Step3Accounting';
