+++ src/routes/DebtTypes/DebtTypes.test.tsx
@@ -4,7 +4,7 @@ import { useNavigate, useSearchParams } from 'react-router-dom';
 import { getDebtPositionTypeWithCount } from '../../api/debtPositionsTypes';
 import DebtTypes from './DebtTypes';
 import { i18nTestSetup } from '../../__tests__/i18nTestSetup';
-import { PageRoutes } from '../../App';
+import { PageRoutes } from '../../routes';
 
 i18nTestSetup({
   'commons.routes.DEBT_TYPES_CATALOG': 'Debt Types Catalog',
