+++ src/routes/DebtTypesCreated/MyOrg/MyOrg.tsx
@@ -17,7 +17,7 @@ import { useStore } from '../../../store/GlobalStore';
 import { STATE } from '../../../store/types';
 import { formatDateTime } from '../../../utils/formatters';
 import { generatePath, useNavigate } from 'react-router-dom';
-import { PageRoutes } from '../../../App';
+import { PageRoutes } from '../../../routes';
 
 type MyOrgProps = {
   codeFilter: string;
