+++ src/routes/UtilityPages/genericError.test.tsx
@@ -1,7 +1,7 @@
 import { describe, it, expect, vi, beforeEach } from 'vitest';
 import { fireEvent, render, screen } from '../../__tests__/renderers';
 import GenericErrorPage from './genericError';
-import { PageRoutes } from '../../App';
+import { PageRoutes } from '../../routes';
 
 const mockNavigate = vi.fn();
 const mockUseLocation = vi.fn();
