+++ src/routes/UtilityPages/success.tsx
@@ -1,6 +1,6 @@
 import { useTranslation } from 'react-i18next';
 import { useNavigate, useLocation } from 'react-router-dom';
-import { PageRoutes } from '../../App';
+import { PageRoutes } from '../../routes';
 import CheckCircleOutlineOutlinedIcon from '@mui/icons-material/CheckCircleOutlineOutlined';
 import { theme } from '@pagopa/mui-italia';
 import { SuccessPageConfig } from '../../models/SuccessPageConfig';
